requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
decorator==4.4.2
distro==1.9.0
edge-tts==7.0.2
ffmpeg==1.4
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.25.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
moviepy==1.0.3
msgpack==1.1.1
multidict==6.6.3
numba==0.61.2
numpy==2.2.4
packaging==25.0
Pillow==9.5.0
platformdirs==4.3.8
pooch==1.8.2
proglog==0.1.11
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pypiwin32==223
pysrt==1.1.2
python-dotenv==1.0.1
pyttsx3==2.99
pywin32==310
pyzmq==27.0.0
requests==2.31.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
srt==3.5.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.2.0
urllib3==2.3.0
websockets==15.0.1
yarl==1.20.1
